Core: Delete duplicate `ReplicaState` enum
